Make sure the HTML is correctly encoded after restoring chunks
This fixes rstudio/rmarkdown#686 and rstudio/bookdown#142. The problem in these cases is that `chunks` is encoded in UTF-8 and contains multibyte characters, but `strval`'s encoding is unknown (native). After gsub(useBytes = TRUE), `strval`'s encoding is still unknown, and on Windows, unknown/native != UTF-8, so when the restored HTML is written out, the multibyte characters are garbled.